Davinci 0.4 initialization code

+++ core/src/main/java/edp/davinci/core/enums/UserOrgRoleEnum.java
@@ -34,7 +34,7 @@ public enum UserOrgRoleEnum {
 
     public static UserOrgRoleEnum roleOf(short role) {
         for (UserOrgRoleEnum userOrgRoleEnum : values()) {
-            if ((short) userOrgRoleEnum.getRole() == role) {
+            if (userOrgRoleEnum.getRole() == role) {
                 return userOrgRoleEnum;
             }
         }
